fix(allocator/split): compute total_size with uintptr_t canary
MEM_splitBlock() accounted for the trailing canary using sizeof(uint32_t),
which is inconsistent with the actual canary type (uintptr_t). On 64-bit
targets this undercounted the block size, misplacing the canary and risking
metadata corruption, false/ missed overflow detections, and free-list issues.

This aligns the block layout across 32/64-bit, fixes the overflow check,
and prevents off-by-N writes/reads when setting/validating the canary.
No ABI impact.

+++ src/libmemalloc.c
@@ -2274,7 +2274,7 @@ static int MEM_splitBlock(mem_allocator_t *const allocator,
 
   aligned_size = ALIGN(req_size);
   total_size
-    = (size_t)(aligned_size + sizeof(block_header_t) + sizeof(uint32_t));
+    = (size_t)(aligned_size + sizeof(block_header_t) + sizeof(uintptr_t));
 
   if (block->size < total_size + MIN_BLOCK_SIZE)
   {
